requirements.txt
numpy~=1.19.5
matplotlib~=3.3.4
scipy~=1.5.3
scikit-learn~=0.24.2
scikit-image~=0.17.2
pytorch>=1.6.0
torchvision>=0.7.0
opencv-python~=4.5.2.52
pyyaml~=5.4.1
tqdm~=4.61.0
yaml~=0.2.5